+++ autoPyTorch/pipeline/components/preprocessing/tabular_preprocessing/feature_preprocessing/RandomKitchenSinks.py
@@ -54,9 +54,10 @@ class RandomKitchenSinks(autoPyTorchFeaturePreprocessingComponent):
             else:
                 log = n_components.log
             n_components = HyperparameterSearchSpace(hyperparameter='n_components',
-                                                     value_range=(floor(int(n_components.value_range[0]) * n_features),
-                                                                  ceil(int(n_components.value_range[1]) * n_features)),
-                                                     default_value=ceil(int(n_components.default_value) * n_features),
+                                                     value_range=(
+                                                         floor(float(n_components.value_range[0]) * n_features),
+                                                         ceil(float(n_components.value_range[1]) * n_features)),
+                                                     default_value=ceil(float(n_components.default_value) * n_features),
                                                      log=log)
         else:
             n_components = HyperparameterSearchSpace(hyperparameter='n_components',
